requirements.txt
plaid-python>=16.0.0
firebase-admin
Flask
flask-cors
python-dotenv
cryptography
joblib
pandas
numpy
requests
scikit-learn
gunicorn